Move dependency metadata code out of Dependency
...and into DependencyMetadataFinders. Dependency is now a simple value
object, and the code that hits Github to find the changelog, etc. is all
in the DependencyMetadataFinders.

The motivation for this is removing the global Github client - until now
the global client has been used all over the codebase, and was pulling
the access token from an environment variable. Obviously this isn't
desirable if you're going to use bump as a library.