Skip gems that aren't in the lockfile
Not all files in the Gemfile will appear in the Gemfile.lock. For
instance, if a gem specifies `platform: [:windows]`, and the
Gemfile.lock is generated on a Linux machine, the gem will be not appear
in the lockfile.

Previously we would blow up in this case, now we just skip over the gem.