Check all top level deps for peer dep requirements
Force yarn to download the manifest for all top level dependencies so
that peer dependency requirements can be checked against the new
version.

Example:

If we want to update react we also need to check react-dom's peer
dependencies to make sure it's a valid update.

If we only pass react to yarn add it won't fetch the manifest for
react-dom and never know about the peer dependency requirement on react.