Npm: Ignore bundled sub-dependencies
Thinking that we shouldn't update bundled sub-dependencies as they have
been bundled into the release at an exact version by a parent using
`bundledDependencies`.

For example, fsevents < 2 bundles node-pre-gyp meaning all it's
sub-dependnecies get bundled into the release tarball at publish time so
you always get the same sub-dependency versions if you re-install a
specific version of fsevents.

Updating the sub-dependency (by deleting the entry) does seem to work
but it gets removed from the bundled set of dependencies and moved top
level resulting in a bunch of package duplication which is pretty
confusing.

To fix this properly we would need to try and update the parent but
this gets a bit crazy if we need to go up several levels which is the
case when updating `tar` here: `chokidar > fsevents > node-pre-gyp >
tar`.