JS: Fix missing previous version when git sha
Fix missing previous version when the dependency already has a git sha
version. This was being cleared out because of an unecessary semver
version class check.

Changing the return value of the resolved previous version to always be
a string which is the default behaviour when a previous version exists
without needing to be resolved from the requirements.