Use project based fixtures fo go_module specs
This PR introduces a project based setups for go_modules, meaning that
every test points to a fixture project, which we then turn into a git
repository for each test.

This sets us up for `go mod tidy` support, in which we'll always assume
that the repository has been checked out.