JS: Handle rate limited npm package requests
When checking if a package is private dependabot makes a request to the
package url, e.g. https://www.npmjs.com/package/express and if this
redirects to the login page it is assumed it is private so we show a
custom PrivateSourceAuthenticationFailure. This request can become rate
limited in which case it returns a 429.