Pip compile: handle unresolvable projects
Check the cause exception when checking resolvability for pip compile.

When rescuing pip compile errors we retry the update on the initial
manifest files to see if the project is unresolvable to begin with, when
checking the exception parse the causing exception for resolvability
errors instead of the new exception that might have a different error
message.