go_modules: don't raise error for go mod tidy
`go mod tidy` makes sure that go.mod matches the source code in a Go
module. This also means that `dependabot` is doing a quassi linting on
the code, which prevents updating versions in a go.mod if the source
code doesn't got build. One particular use case is when a project is
missing generated files. Sometimes people don't submit and commit the
generated files of a code generator (i.e: `protobuf`) and do it in
pre-script in the CI. Because `dependabot` is not aware of this
pre-script, it'll run `go mod tidy` and fail for the user.

We should silently continue if `go mod tidy` fails and not raise an
error. This will allow dependabot still to run `go mod tidy`, but will
not block users for these kind of edge cases.