npm: fix npm workspace bug
Fixes one of the bugs discovered in this update:
https://github.com/AviVahl/ts-tools/pull/152

Dependabot will no longer attempt to add the nested dependency to the
root level lockfile when doing updates to npm workspace projects.

This is done by checking if the dependency being updated is defined in a
`package.json` in the same folder as the lockfile being updated. If it's
a nested package we run `npm install` without any package@version
arguments letting npm update the nested package references which have
already been updated in the package.json by the `PackageJsonUpdater`.

Additionally, I've added a workaround to stop adding the `name`
attribute to the lockfile `packages.""` section when running `npm
install` with `--package-lock-only`. This seems to be a bug with npm but
without this fix dependabot will always add this attribute and it will
most likely never exist in a users repo so adding confusion about why
it's added.

TODO
- [ ] Add more specs covering these cases