Add option to profile dry-run using Stackprof
This makes it easy to figure out performance bottlenecks by analyzing
where time is spent in a run. The trace can be visualized by running
`stackprof --d3-flamegraph tmp/stackprof-2021-02-15-13:00.dump > tmp/flamegraph.html`
and viewing that in a browser. For some reason the built in viewer did
not work on my machine, but the d3-generated flamegraph has been clear
enough for me.