Pull Requests: Sanitize team mentions
When we pull in changelogs and commits etc we sanitize user mentions
(`@jurre`), by wrapping them in a codeblock and inserting a zero-width
space between the `@` and the username, and wrapping that in a link to
the profile, to prevent users from getting notifications from Dependabot PRs.

Something that we didn't cover before yet were team-mentions, e.q:
@dependabot/reviewers. These are slightly tricky in the sense that there
are ecosystems (npm for example) that have package names that follow the
same pattern. There is no way to distinguish between a package name and
a team mention, so we treat them the same and sanitize them but do not
attempt to link, as this is expected to lead to a lot of false
positives.

Fixes https://github.com/dependabot/dependabot-core/issues/3378