Bundler 2: No longer bump yanked gems when updating dependency
Prior to bundler 2.2.17, updating any dependency while a locked version
of another dependency was yanked, would result in our `LockfileUpdater`
native helper to encounter a `Bundler::GemNotFound` error, and to
resolve this, it would add the offending yanked gem to the
`dependencies_to_unlock` and install the latest allowed version of that
gem.

In 2.2.17 though, _something_ has changed, and this error is no longer
raised.

It seems like this change is actually desirable as we keep the PR as
small as possible and don't update unrelated gems in the process, but
I've not been able to find the exact change in bundler that causes this
new behavior.