Bundler tests: Use git sources for organizations that we own
Bundler git dependencies have the possibility to execute code during the
installation phase, to ensure this does not happen in our test suite,
this updates any tests that pull in a git dependency to reference an
organization/account that we own.

This has the added benefit of protecting us against those repositories
being renamed, moved or deleted.