Clarify support policies in README and Issue template
We tend to get a lot of questions here about the Dependabot service that
we operate at GitHub. Many of those questions are much easier to resolve
when going through Support, as they have much better tooling and
processes to follow up on those sorts of questions.

This attempts to clarify that in the README and issue template.