Add test for PoetryFileUpdater with a python_index with auth details
This was tested with a local instance of pypiserver with an empty packages/
directory, which meant it was just proxying calls to pypi.org.

The credentials used in testing were set up using the pypiserver instructions at
https://pypi.org/project/pypiserver/#apache-like-authentication-htpasswd.

This might be feasible for Dependabot's own testing, as you could run this
as part of the setup, but there are probably better alternatives.