github_actions: Add support for Composite Actions
Composite Actions follow a very similar structure to the "normal"
GitHub Actions, with the exception that they must be located in a
file called action.yml (or .yaml) in the root of the repository.

Because of this similarity the file_parser and the file_updater does
not need any tweaking, and it is only the file_fetcher that needs to
be able to search beyond the .github/workflows folder.

Since GitHub only looks for a single file in the root directory of
the repository we can limit the expansion to the search to the same
strict parameters so we don't accidentally find a lot of other stuff.

Resolves #4178