Modify ENV in a friendlier way during specs
Right now, if one of these specs start failing, we we'll get a very
confusing error like the following:

```
Failures:

  1) Dependabot::Bundler::NativeHelpers.run_bundler_subprocess with DEPENDABOT_NATIVE_HELPERS_PATH not set uses the full path to the uninstalled run.rb command
     Failure/Error:
       expect(Dependabot::SharedHelpers).
         to have_received(:run_helper_subprocess).
         with(
           command: "bundle exec ruby #{File.expand_path('../../../v2/run.rb', __dir__)}",
           function: "noop",
           args: [],
           env: anything
         )

       {"SHELL"=>"/bin/zsh", (... rest of the env...), "GEM_HOME"=>"/Users/deivid/.rbenv/versions/3.1.1/lib/ruby/gems/3.1.0"} received :[] with unexpected arguments
         expected: ("DEPENDABOT_NATIVE_HELPERS_PATH")
              got: ("COLUMNS")
        Please stub a default value first if message might be received with other args as well.
     # ./spec/dependabot/bundler/native_helpers_spec.rb:100:in `block (4 levels) in <top (required)>'
     # ./spec/spec_helper.rb:54:in `block (2 levels) in <top (required)>'
     # /Users/deivid/Code/dependabot/dependabot-core/common/spec/spec_helper.rb:49:in `block (2 levels) in <top (required)>'
```

I think this is because RSpec needs ENV to build its error messages, but
it's still stubbed when error messages are built.

We could add `allow(ENV).to receive(:[])` to allow calls with other
arguments on top of the current stub. But I think it would be a bit
confusing to figure out why that's needed.

I think it's better to not mock ENV and instead change and restore the
appropriate variable as needed.