Pip: Do not raise PathDependenciesNotReachable for missing setup.py
Previously, Dependabot would assume that for every requirements.txt with
a self-reference ('`-e .`), we should expect a `setup.py` to be present. The
self-reference is considered a path dependency that needs a setup.py.

However, nowadays it's valid to have a requirements.txt without a
setup.py. This change accounts for that by still attempting to pull in a
setup.py if it exists, but to not raise and error when it doesn't.