Reference latest blinker version in spec that references public package
This is unrelated to other changes in this PR, but blinker released a
new version today, and this breaks one of our tests.

We should replace it with a reference to something we own, or lock the
constraints so that we don't run into this, but this get's us green for
now at least.