Make failures in Bundler 1 native helper specs actually fail CI
This is a fun one.

Consider the following script

```bash
#!/bin/bash

set -e

true

if [[ "$SUITE_NAME" == "bundler1" ]]; then
  cd helpers/v1 \
    && false \
    && cd -
fi

if [[ "$SUITE_NAME" == "bundler2" ]]; then
  cd helpers/v2 \
    && false \
    && cd -
fi
```

Believe or not, this is what you get when you run it

```
$ SUITE_NAME=bundler1 ./script/ci-test-dummy; echo $?
=> 0

$ SUITE_NAME=bundler2 ./script/ci-test-dummy; echo $?
=> 1
```

You can read more about this here:
http://mywiki.wooledge.org/BashFAQ/105, but the conclusion is that `set
-e` is brittle.

Yet I think adding `|| exit 1` to all commands is painful.

As an alternative, I added a full if-else construct that works as
expected, with a note on top. I'm not fully sold on this solution, but
it's simple and it also lets the script default to Bundler 2 when run
without `SUITE_NAME` set, which seems handy.