Use commits/tags consistently
I'm a little torn on this, but it's been pitched to me that we should
upgrade to a commit if a user has pinned their Action to a commit
already. My hesitation has to do with a lack of transparency between the
version/tag to which the commit belongs. However, that transparency is
always lacking when pinning to a commit.