Fix updating to tags with a branch with same name
When detecting whether a workflow file is pinned, we were assuming it's
not pinned if there's a branch that matches the ref specified in the
workflow file.

However, if there's also a tag with the same name, that should probably take
priority, since that's what users will expect in this case.